requirements.txt
Django==4.2.4
djangorestframework==3.15.2
requests==2.32.3
